Restructure public api unit tests
