+++ scripts/hooks/pre-commit
@@ -31,9 +31,6 @@ EOF
     exit 1
 fi
 
-# If there are whitespace errors, print the offending file names and fail.
-git diff-index --check --cached $against -- || exit 2
-
 # Reformat C/C++ files that are part of the commit
 for file in $(git diff --cached --name-only | grep -E '\.(c|h|cpp|hpp)$'); do
     echo "Formatting $file"
